Move etcd client to own package.
